# cmd/swag/docs/swagger.yaml
basePath: /
definitions:
...
  openai.ChatCompletionRequest:
    properties:
      max_tokens:
        type: integer
      messages:
        items:
          $ref: '#/definitions/openai.Message'
        type: array
      model:
        type: string
      stream:
        type: boolean
      temperature:
        type: number
    type: object
  openai.ChatCompletionResponse:
    properties:
      choices:
        items:
          $ref: '#/definitions/openai.Choice'
        type: array
      created:
        type: integer
      id:
        type: string
      model:
        type: string
      object:
        type: string
      usage:
        $ref: '#/definitions/openai.Usage'
    type: object
  openai.Choice:
    properties:
      finish_reason:
        type: string
      index:
        type: integer
      message:
        $ref: '#/definitions/openai.Message'
    type: object
  openai.Error:
    properties:
      code:
        type: string
      message:
        type: string
      type:
        type: string
    type: object
  openai.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/openai.Error'
    type: object
  openai.Message:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
...
  openai.Usage:
    properties:
      completion_tokens:
        type: integer
      prompt_tokens:
        type: integer
      total_tokens:
        type: integer
    type: object
host: localhost:3000
info:
  contact: {}
  description: "\U0001F680 High-performance WebAI-to-API gateway. Seamlessly bridge
    Google Gemini into standardized OpenAI, Anthropic (Claude), and Google Native
    REST APIs."
  title: AI Bridges API
  version: "1.0"
paths:
  /v1/chat/completions:
    post:
      consumes:
      - application/json
      description: Accepts requests in OpenAI format
      parameters:
      - description: Chat request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/openai.ChatCompletionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.ChatCompletionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/openai.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/openai.ErrorResponse'
      summary: OpenAI-compatible chat completions
      tags:
      - OpenAI Compatible
...
swagger: "2.0"